+++ huginn-proxy-lib/tests/security/mod.rs
@@ -1,3 +1,5 @@
+pub mod headers;
+pub mod ip_filter;
 pub mod rate_limit;
 pub mod rate_limit_estimator;
 pub mod rate_limit_limiter;
